feat: improve experience when working with disabled rails components
Automatically mocks configuration for disabled Rails components, such as ActiveJob, and makes the
experience of disabling it easier, as we no longer need to comment out the configuration for it in
the `environments/<env>.rb` file. This is super useful when you want to start a new project and
aren't sure you're going to use a component or not, but would like to leave its configuration intact
in case you change your mind.

With this enabled (default), it will not raise exceptions when configuring a specific framework
configuration such as `config.active_job.log_tags` when ActiveJob is commented out on
`config/application.rb`